Add more features in Day 9
- Adding Character Scriptable Object
- Integrate Character Scriptable Object with other scritps
- Add Death Animation to Player and Enemy
- Fixing bug by Spell Object SO

+++ Assets/Script/Spell/SpellSO.cs
@@ -11,15 +11,15 @@ namespace Survival.Spell
     public class SpellSO : ScriptableObject 
     {
         [SerializeField] SpellConfig spell;
-        [SerializeField] float damageSpell;
-        [SerializeField] float manaConsumed;
+        [SerializeField] int damageSpell;
+        [SerializeField] int manaConsumed;
         
 
-        public float GetDamageSpell()
+        public int GetDamageSpell()
         {
             return damageSpell;
         }
-        public float GetManaConsumed()
+        public int GetManaConsumed()
         {
             return manaConsumed;
         }
